WIP Relative color syntax (RCS) (without class)
https://csswg.sesse.net/css-color-5/#relative-colors
go/rcs-chromium

This CL launches RCS behind a flag. Tests are added to the
virtual/stable test suite to ensure that it is not accidentally
launched on stable.

With RCS, color channels can now be math expressions with possible
"channel keyword" substitutions. To accomplish this, an optional
hash map of {CSSValueID, double} has been added to necessary
CSSMathExpressionNode functions. When the parser encounters one of the
valid CSSValueIDs it simply returns the associated double instead of
failing. This color_channel_keyword_values HashMap could be given a more
generic name if it could be useful for other features in the future.

With this CL, chrome now passes 94% of 1922 tests for valid/computed
relative colors. The remaining failures are mostly problems with the
tests themselves, such as the following:
https://github.com/w3c/csswg-drafts/issues/9094
This is also the case for the new failures in
color-invalid-relative-color.html

Some basic incorrect tests were changed here. For example, lightness
is always clamped to [0, 100]. A fuzzy test for color channels was
added to the "valid" test harness, exactly like what already exists in
the "computed" version.

There are two major TODOs:
  - calc expressions that mix types are not currently handled. i.e. lab(from magenta calc(l - 20%) a b).
  - currentcolor as an origin color does not work.

The first is hopefully easy to resolve in a follow up CL.
The second is not currently handled in any browser and requires some
thought, as with currentcolor the values of the channel keywords will
not be known at parse time. A possible solution is to simply pass in
any random double value for the channel keywords and verify that the
expression does indeed parse, but then save that expression and actually
parse it at computed value time.

Bug: 1447327
Change-Id: I8eb3bc780deb700aa4741ad071b05287c0c0deb3